Extract FolderService to Core for folder CRUD operations
Create stateless FolderService with methods for finding, moving, and
copying folder definitions. Refactor code-behind to delegate folder
lookups and mutations, removing ~70 lines of recursive folder logic.
